Add files via upload
The goal of this script is to back up important files on your system, saving them to an external directory or drive. 
The script can also compress the files to save space. 

Requirements: Have permissions to read the files and write to the destination directory. Have enough space on the target disk. Know the path of the files/directories to be backed up. Script Steps: Select directories to back up (e.g., /home/user/Documents, /etc/, etc.). Create a destination directory where backups will be stored (e.g., /backups/). Compress the files (using tar) to save space. Save the compressed files with a timestamp (date and time) to be able to differentiate them.